refactor v1.0.1
